fix(front): renames creative commons svg and import
possibly a conflict between HTML comment and text content
