+++ front/src/lib/components/icons/CreativeCommons.svelte
